* Major refactoring of the filter engine to improve compile times as well as stream line implementation process. * Added where filters to new CheckFiles (removed deprecated CheckFile command). * New concept for the CheckTaskSchedule command (yes it will be documented this time) * A lot of changes in general to improve the where filters. Also this is largly untested as of yet so be carefull if you wanna build...

+++ include/simple_timer.hpp
@@ -23,7 +23,7 @@ public:
 		start_time = getFT();
 	}
 	unsigned long long stop() {
-		unsigned int  ret = getFT() - start_time;
+		unsigned int ret = static_cast<unsigned int>(getFT() - start_time);
 		start();
 		return ret/1000;
 	}
